Move the graphics pipeline tests to separate module
